move test for changes in PR 2185
